move of profiles views and urls into dedicated app
