+++ oc_lettings_site/urls.py
@@ -6,7 +6,6 @@ from . import views
 urlpatterns = [
     path('', views.index, name='index'),
     path('letting/',include('lettings.urls')),
-    path('profiles/', views.profiles_index, name='profiles_index'),
-    path('profiles/<str:username>/', views.profile, name='profile'),
+    path('profiles/',include('profiles.urls')),
     path('admin/', admin.site.urls),
 ]
